refactor(bpmn-modeler): move diff rendering primitives to /viewer subpath (#1439)
DiffViewer, DiffLegend, DiffNavigator and DiffPaneCoordinator now ship from
@miragon/bpmn-modeler/viewer; the root entry keeps deprecated aliases. The
neutral marker/legend CSS splits into diffView.css, shared by viewer.css and
styles.css. The viewer purity gate (check-viewer-pure-entry.mjs + the
architecture-spec rule) is retired: the viewer surface will keep growing
(i18n via DiffLegend now, readonly panel in #1443). ADR 0014 amended.
