+++ apps/demo-webapp/bpmn/diff.html
@@ -3,10 +3,11 @@
     <head>
         <meta charset="UTF-8" />
         <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
-        <!-- The engine + theme CSS ships via the `@miragon/bpmn-modeler` side
-             effects imported from diff.ts (Vite serves it in dev; a raw <link>
-             to the package source would escape the dev-server root and 404).
-             Both panes theme off the shared demo header's `data-bpmn-theme`. -->
+        <!-- The lean `@miragon/bpmn-modeler/viewer.css` (base diagram CSS + diff
+             markers/legend) is imported from diff.ts so Vite processes it in dev
+             and build (a raw <link> to the package source would escape the
+             dev-server root and 404). Both panes theme off the shared demo
+             header's `data-bpmn-theme`. -->
         <meta content="width=device-width, initial-scale=1.0" name="viewport" />
         <title>Miragon Modeler — Two-Pane Diff Demo</title>
         <style>
