+++ packages/bpmn-modeler/src/viewer/diff/DiffLegend.spec.ts
