+++ packages/bpmn-modeler/src/viewer/diff/DiffLegend.ts
