+++ packages/bpmn-modeler/src/viewer/diff/DiffNavigator.spec.ts
