+++ packages/bpmn-modeler/src/viewer/diff/DiffNavigator.ts
